ref(minidump): Do not store minidumps before processing
